# Conflicts:
#	wifi4eu-dgconn/wifi4eu-dgconn-service/src/main/java/wifi4eu/wifi4eu/util/ScheduledTasks.java
#	wifi4eu-dgconn/wifi4eu-dgconn-web/src/main/angular/src/app/+dgconn-portal/+applicant-registrations/+details/applicant-registrations-details.component.html
#	wifi4eu-scripts/releases/2.1.0-RC2/update-db.sql

+++ wifi4eu-portal/wifi4eu-portal-service/src/main/java/wifi4eu/wifi4eu/service/registration/legal_files/LegalFilesService.java
@@ -54,7 +54,7 @@ public class LegalFilesService {
 	public boolean hasUserPermissionForLegalFile (Integer registrationId, Integer userId, Integer fileId){
 		if(registrationService.checkUserWithRegistration(registrationId, userId)){
             LegalFile legalFile = legalFilesRepository.findOne(fileId);
-            if (legalFile != null && (userId == legalFile.getUserId() || legalFile.getFileType() == FileTypes.LEGALFILE1.getValue() || legalFile
+            if (legalFile != null && (userId == legalFile.getUserId().intValue() || legalFile.getFileType() == FileTypes.LEGALFILE1.getValue() || legalFile
                     .getFileType() == FileTypes.LEGALFILE3.getValue())) {
                 return true;
             }
